+++ docs/CHANGELOG.md
@@ -21,6 +21,7 @@ For git tags, official releases are marked like "v1.0.0", while unofficial ones 
 ### Changed
 
 * Prettier UI to navigate between text creation/edition/archive pages.
+* When reading text, you can click on "Show All" and "Translations" to click the checkbox.
 * The long text import page looks a bit nicer.
 
 ### Fixed